+++ scripts/com.r2studio.TsumBeta/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the TsumBeta script will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [v80] - 2025-12-19
+
+### Fixed
+- Current JP's princess pickup capsule changed colors which broke the auto buy functionality.
+
+
 ## [v79] - 2025-12-10
 
 ### Fixed